Add full player grouping functionality
The only situation this has been properly tested in is with a single player joining a pre-existing 2-player group, but logically it should handle an infinite number of players. The final step in player grouping is to handle full group merges, where two group's center's get within range of each other, which should cause them to merge into a singular group. Theoretically the best way to approach this is to create a new group with all players involved and delete the previous two groups.