Rewrite HTTP request/response system for voice groups
Requests are now event based. The database will no longer need to reconstruct what happened on the server through comparing values to requests, it will simply execute what needs to be done based on the request. Groups now merge on getting close enough to each other, and the database handles it flawlessly. On starting a server, the database will make sure to clean up the leftover voice channels from the last session. This is to make sure reloading servers works on discord side, as well as to handle Bedrock server crashes properly. More testing is required for edge cases, but this should be the near final behaviours for the addon.